requirements.txt
einops
timm
tensorboard
gensim
torch-summary